Add 'now' command.
Allow users to query whether it is day or night right now according to
either sunrise/sunset or civil dawn/dusk.

It would be nice if there was an enum to represent the various
Natural/Civil/Nautical/Astronomical types, which would make it easier
to provide multiple options rather than just -c/--civil, but this will
serve many needs for now.